разделение MapActivity на классы
